Move UserSignatureBuilder from Reko.Analysis to Reko.Core.CLanguage
