+++ src/Decompiler/Scanning/Scanner.cs
@@ -18,8 +18,8 @@
  */
 #endregion
 
-using Reko.Analysis;
 using Reko.Core;
+using UserSignatureBuilder = Reko.Core.CLanguage.UserSignatureBuilder;
 using Reko.Core.Code;
 using Reko.Core.Configuration;
 using Reko.Core.Expressions;
