+++ src/UnitTests/Analysis/UserSignatureBuilderTests.cs
@@ -20,7 +20,6 @@
 
 using Moq;
 using NUnit.Framework;
-using Reko.Analysis;
 using Reko.Core;
 using Reko.Core.CLanguage;
 using Reko.Core.Code;
